tests/wpt/web-platform-tests/css/css-shadow-parts/invalidation-change-part-name.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Shadow Parts - Invalidation Change Part Name</title>
    <meta href="mailto:fergal@chromium.org" rel="author" title="Fergal Daly">
    <link href="http://www.google.com/" rel="author" title="Google">
    <link href="https://drafts.csswg.org/css-shadow-parts/" rel="help">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/shadow-helper.js"></script>
  </head>
  <body>
    <style>#c-e::part(partp) { color: red; }</style>
    <script>installCustomElement("custom-element", "custom-element-template");</script>
    <template id="custom-element-template">
      <style>span { color: green; }</style>
      <span id="part" part="partp">This text</span>
    </template>
    The following text should be green:
    <div><custom-element id="c-e"></custom-element></div>
    <script type="text/javascript">
      "use strict";
      test(function() {
        var part = getElementByShadowIds(document, ["c-e", "part"]);
        var before = window.getComputedStyle(part).color;
        part.setAttribute("part", "new-partp");
        var after = window.getComputedStyle(part).color;
        assert_not_equals(before, after);
      }, "Part in selected host changed color");
    </script>
  </body>
</html>
